requirements.txt
bs4
more-itertools
python-dotenv
pandas
pytest
requests